長くなってきたアプリケーションエントリーポイント内記述を別ファイル化した
